% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_collider.R
\name{analysis_collider}
\alias{analysis_collider}
\title{analysis_collider}
\usage{
analysis_collider(data)
}
\arguments{
\item{data}{A nested data frame containing the processed data.}
}
\value{
Results of analysis.
}
\description{
Fit a collider configural model (M ~ X + Y; Y ~ X) to the data using linear regression and a Maximum Likelihood estimator. Extract the beta estimates, their 95\% CIs and p values.
}
