% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_confounder.R
\name{analysis_confounder}
\alias{analysis_confounder}
\title{analysis_confounder}
\usage{
analysis_confounder(data)
}
\arguments{
\item{data}{A nested data frame containing the processed data.}
}
\description{
Fit a confounder configural model (Y ~ X + M) to the data using linear regression and a Maximum Likelihood estimator. Extract the beta estimates, their 95\% CIs and p values.
}
