% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_mediation.R
\name{analysis_mediation}
\alias{analysis_mediation}
\title{analysis_mediation}
\usage{
analysis_mediation(data)
}
\arguments{
\item{data}{A nested data frame containing the processed data.}
}
\description{
Fit a mediation configural model (Y ~ X + M; M ~ X) to the data using linear regression and a Maximum Likelihood estimator. Extract the beta estimates, their 95\% CIs and p values.
}
