% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_regression_ols.R
\name{analysis_regression_ols}
\alias{analysis_regression_ols}
\title{analysis_regression_ols}
\usage{
analysis_regression_ols(data)
}
\arguments{
\item{data}{A nested data frame containing the processed data.}
}
\description{
Fit an Ordinary Least Squares linear regression to the data using the model Y ~ X and extract the beta estimate, its 95\% CIs and p values.
}
\details{
One use case for this package is to compare how differently specified models fit to a given data generating model (e.g., generating collider model data and fitting simple regression, mediation, collider, and confound models to it). Those other models will be fit using lavaan (and therefore maximum-likelihood estimation over OLS). In order to compare like with like, it may therefore be more appropriate to use analysis_regression_ml() for such a simulation. Other use cases may prefer a simple OLS regression. I therefore include both OLS/lm() and ML/lavaan::sem() implementations of bivariate regression.
}
