% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_ttest.R
\name{analysis_ttest}
\alias{analysis_ttest}
\title{analysis_ttest}
\usage{
analysis_ttest(data)
}
\arguments{
\item{data}{A nested data frame containing the processed data.}
}
\value{
Results of analysis.
}
\description{
Fit a t test to the data using the model Y ~ X and extract the estimated mean difference, the effects, 95\% CIs and p values.
}
\examples{
# population model
population_model <-
  create_population_model_with_random_item_loadings(
    model_specification = "Y_latent ~ 0.2*X_latent",
    item_loading_min_y = 0.5,
    item_loading_max_y = 0.9,
    n_indicators_y = 8,
  )

# run a simulation
results <- 
  generate_data(pop_model_label = "ttest indicators",
                pop_model = population_model, 
                factorial_design = TRUE,
                #n = 50, 
                n_mean = 100,
                n_sd = 25,
                iterations = 100) |>
  data_preprocessing(method = convert_to_likert) |>
  data_processing(method = use_latent_scores) |>
  fit_model(analysis = analysis_ttest)

}
