% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_to_likert.R
\name{convert_to_likert}
\alias{convert_to_likert}
\title{convert_to_likert}
\usage{
convert_to_likert(data)
}
\arguments{
\item{data}{A nested data frame.}
}
\value{
A nested data frame of data_raw + new preprocessed data column.
}
\description{
Convert continuous raw data to interval Likert data.
}
\examples{
# specify population model
population_model <-
  create_population_model_with_static_item_loadings(
    model_specification = 
       "Y_latent ~ 0.0*X_latent + 0.0*M_latent; X_latent ~~ 0.0*M_latent",
    item_loading_y = 0.9,
    item_loading_x = 0.7,
    item_loading_m = 0.5,
    n_indicators_y = 10,
    n_indicators_x = 10,
    n_indicators_m = 10
  )

generated_data <-
  # generate data
  generate_data(pop_model_label = "covariate indicators",
                pop_model = population_model,
                n = 100,
                iterations = 2) |>
  # convert to likert
  data_preprocessing(method = convert_to_likert)

}
