% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/create_population_model_with_random_item_loadings.R
\name{create_population_model_with_random_item_loadings}
\alias{create_population_model_with_random_item_loadings}
\title{create_population_model_with_random_item_loadings}
\usage{
create_population_model_with_random_item_loadings(
  model_specification,
  item_loading_min_y = 0.99,
  item_loading_max_y = 0.99,
  item_loading_min_x = 0.99,
  item_loading_max_x = 0.99,
  item_loading_min_m = 0.99,
  item_loading_max_m = 0.99,
  n_indicators_y = 1,
  n_indicators_x = 1,
  n_indicators_m = 1
)
}
\arguments{
\item{model_specification}{string lavaan model specification}

\item{item_loading_min_y}{minimum factor loading for variable Y}

\item{item_loading_max_y}{maximum factor loading for variable Y}

\item{item_loading_min_x}{minimum factor loading for variable X}

\item{item_loading_max_x}{maximum factor loading for variable X}

\item{item_loading_min_m}{minimum factor loading for variable M}

\item{item_loading_max_m}{maximum factor loading for variable M}

\item{n_indicators_y}{number of indicators for varible Y}

\item{n_indicators_x}{number of indicators for varible X}

\item{n_indicators_m}{number of indicators for varible M}
}
\value{
A nested data frame of data_raw + new scored data column.
}
\description{
Generate lavaan syntax for latent variables Y, X and M, and paste together with string specifying regression models among them.
}
\details{
Instead of using the same item loading for all indicators, define a min and max for each scale and sample from a uniform distribution between these values.
}
\examples{
create_population_model_with_random_item_loadings(
  model_specification = "Y ~ 0.0*X + 0.0*M; X ~~ 0.0*M",
  item_loading_min_y = 0.4,
  item_loading_max_y = 0.9,
  item_loading_min_x = 0.4,
  item_loading_max_x = 0.9,
  item_loading_min_m = 0.4,
  item_loading_max_m = 0.9,
  n_indicators_y = 10,
  n_indicators_x = 10,
  n_indicators_m = 10
)

}
