% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/create_population_model_with_static_item_loadings.R
\name{create_population_model_with_static_item_loadings}
\alias{create_population_model_with_static_item_loadings}
\title{create_population_model_with_static_item_loadings}
\usage{
create_population_model_with_static_item_loadings(
  model_specification,
  item_loading_y = 0.99,
  item_loading_x = 0.99,
  item_loading_m = 0.99,
  n_indicators_y = 1,
  n_indicators_x = 1,
  n_indicators_m = 1
)
}
\arguments{
\item{model_specification}{string lavaan model specification}

\item{item_loading_y}{factor loading for variable Y}

\item{item_loading_x}{factor loading for variable X}

\item{item_loading_m}{factor loading for variable M}

\item{n_indicators_y}{number of indicators for varible Y}

\item{n_indicators_x}{number of indicators for varible X}

\item{n_indicators_m}{number of indicators for varible M}
}
\value{
A nested data frame of data_raw + new scored data column.
}
\description{
Generate lavaan syntax for latent variables Y, X and M, and paste together with string specifying regression models among them.
}
\examples{
create_population_model_with_static_item_loadings(
  model_specification = "Y ~ 0.0*X + 0.0*M; X ~~ 0.0*M",
  item_loading_y = 0.9,
  item_loading_x = 0.7,
  item_loading_m = 0.5,
  n_indicators_y = 10,
  n_indicators_x = 10,
  n_indicators_m = 10
)
}
