% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preprocessing.R
\name{data_preprocessing}
\alias{data_preprocessing}
\title{data_preprocessing}
\usage{
data_preprocessing(nested_data, method)
}
\arguments{
\item{nested_data}{A nested data frame containing a column named data_raw, which is nested data frames of raw data, produced by the data generation functions.}

\item{method}{the preprocessing function to be applied to data_raw, e.g., \code{"convert_to_likert()"}.}
}
\value{
A nested data frame of data_raw + new preprocessed data column.
}
\description{
Apply a data processing function to the raw data.
}
