% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_model.R
\name{fit_model}
\alias{fit_model}
\title{fit_model}
\usage{
fit_model(nested_data, analysis)
}
\arguments{
\item{nested_data}{A nested data frame containing a column named data_raw, which is nested data frames of raw data, produced by the data generation functions.}

\item{analysis}{the analysis function to be applied to data_raw, e.g., \code{"analysis_ttest()"}.}
}
\value{
A nested data frame of data_processed + new fit column.
}
\description{
Apply an analysis function to the processed data such as \code{"analysis_ttest()"}.
}
