% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_data.R
\name{generate_data}
\alias{generate_data}
\title{generate_data}
\usage{
generate_data(
  pop_model_label,
  pop_model,
  iterations,
  n,
  n_mean,
  n_sd,
  factorial_design = FALSE
)
}
\arguments{
\item{pop_model_label}{string label for the population model}

\item{pop_model}{string lavaan specification for the population model}

\item{iterations}{number of iterations aka simulated studies}

\item{n}{sample size per iteration (simulated study). Either n or both n_mean and n_sd must be specified but not both.}

\item{n_mean}{integer mean sample size per iteration (simulated study). Either n or both n_mean and n_sd must be specified but not both.}

\item{n_sd}{integer SD sample size per iteration (simulated study). Either n or both n_mean and n_sd must be specified but not both.}

\item{factorial_design}{boolean indicating if data should be generated for a single group cross sectional design or a two group factorial design.}
}
\value{
A nested data frame of data_raw + new scored data column.
}
\description{
Generate simulated data using a given population model, n iterations, and sample size.
}
\examples{
# Example 1: Population is a covariate regression model. 
# Each sample from the population is of equal size.
population_model <-
  create_population_model_with_static_item_loadings(
    model_specification =
      "Y_latent ~ 0.5*X_latent + 0.0*M_latent;
       X_latent ~~ 0.0*M_latent",
    item_loading_y = 0.8,
    item_loading_x = 0.8,
    item_loading_m = 0.8,
    n_indicators_y = 10,
    n_indicators_x = 10,
    n_indicators_m = 10
  )

simulated_data <-
  generate_data(pop_model_label = "covariate indicators",
                pop_model = population_model,
                n = 100,
                iterations = 50)
                
# Example 2: Population is a between-groups difference in means. 
# Studies vary in their sample sizes.
population_model <-
  create_population_model_with_random_item_loadings(
    model_specification = "Y_latent ~ 0.25*X_latent",
    item_loading_min_y = 0.5,
    item_loading_max_y = 0.9,
    n_indicators_y = 8,
  )

# run a simulation
results <- 
  generate_data(pop_model_label = "ttest indicators",
                pop_model = population_model, 
                factorial_design = TRUE,
                n_mean = 100,
                n_sd = 25,
                iterations = 25)

}
