% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_data_cohens_d.R
\name{generate_data_cohens_d}
\alias{generate_data_cohens_d}
\title{generate_data_cohens_d}
\usage{
generate_data_cohens_d(
  pop_model_label,
  n1,
  n2,
  m1,
  m2,
  sd1,
  sd2,
  cohens_d = NULL,
  iterations
)
}
\arguments{
\item{pop_model_label}{string label for the population model, e.g., "confounder indicators" for a confounder causal model with data for individual indicators (e.g., self report items) as well as the latent scores (e.g., sum scores)}

\item{n1}{sample size for group 1.}

\item{n2}{sample size for group 2.}

\item{m1}{population mean (mu) for group 1.}

\item{m2}{population mean (mu) for group 2.}

\item{sd1}{population SD (sigma) for group 1.}

\item{sd2}{population SD (sigma) for group 2.}

\item{cohens_d}{the population standardized mean difference effect size (Cohen's d). You must only supply values for either (a) m1, m2, sd1, and sd2 OR (b) cohens_d but not both.}

\item{iterations}{number of iterations aka simulated studies.}

\item{pop_model}{string specifying population model using lavaan notation.}
}
\value{
A nested data frame.
}
\description{
Generate simulated data using a given population model, n iterations, and sample size.
}
\details{
To generate both item level data and latent scores, use a helper function (e.g., create_population_model_with_static_item_loadings()
or create_population_model_with_random_item_loadings()) create the model specification which must include the assignment of indicators via "=~".
To generate only the latent variables, specify only the regressions (e.g., "Y ~ 0.5\emph{X + 0.0}M"). See the vignettes for examples.

Note that for the later processing, hacking, and analysis functions to work, generate_data() must produce data with the correct variable names:
To generate item level data and latent variables, you must specify the population model to include variables "Y_latent", "X_latent", and (for some) "M_latent".
To generate just the latent variables/scored data, you must specify the population model to include the variables "Y", "X", and (for some) "M" (i.e., no "_latent").
}
