% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_lavaan_syntax_latent.R
\name{generate_lavaan_syntax_latent}
\alias{generate_lavaan_syntax_latent}
\title{generate_lavaan_syntax_latent}
\usage{
generate_lavaan_syntax_latent(variable_name, item_loading, n_indicators)
}
\arguments{
\item{variable_name}{name of variable(s) as it appears in the data frame}

\item{item_loading}{factor loading(s)}

\item{n_indicators}{number of indicators per variable}
}
\value{
A nested data frame of data_raw + new scored data column.
}
\description{
Generate lavaan syntax for latent variables.
}
