% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heterogeneity_metrics_for_forest.R
\name{heterogeneity_metrics_for_forest}
\alias{heterogeneity_metrics_for_forest}
\title{heterogeneity_metrics_for_forest}
\usage{
heterogeneity_metrics_for_forest(fit)
}
\arguments{
\item{fit}{An rma object returned by (\code{"metafor::rma()"} or (\code{"metafor::rma.mv()"}, i.e. a fitted meta analysis model.}
}
\value{
A string reporting the heterogeneity metrics that can be reported in a base R forest plot.
}
\description{
Create a string reporting tau2, I2, and H2 from a meta-analysis fitted in metafor.
}
