% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_covariance.R
\name{learn_covariance}
\alias{learn_covariance}
\title{Create a matrix of covariance}
\usage{
learn_covariance(df, method = "lm")
}
\arguments{
\item{df}{Dataframe containing the real dataset.}

\item{method}{Method to used to estimate the eigenvalues of the matrix,
default='lm'. See details.}
}
\value{
A matrix with the estimated covariance.
}
\description{
This function estimates the covariance matrix of a real dataset.
}
\details{
Two methods are available for the estimation of the eigenvalues of
the matrix. For \code{lm}, we fit a linear model on the log of the first
eigenvalues on the log of their rank. For \code{min}, we add the minimum of the
estimated eigenvalues to the complete set of eigenvalues to ensure that they
are all positives.
}
\examples{
\dontrun{
if(interactive()){
 attach(powerconsumption)

 # Using 'lm'
 cov <- learn_covariance(powerconsumption, 'lm')

 # Using 'min'
 cov <- learn_covariance(powerconsumption, 'min')
 }
}
}
\seealso{
\code{\link[stats]{smooth.spline}},
\code{\link[stats]{predict}},
\code{\link[fdapace]{Lwls2D}}
}
