% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_noise.R
\name{learn_noise}
\alias{learn_noise}
\title{FUNCTION_TITLE}
\usage{
learn_noise(df)
}
\arguments{
\item{df}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 attach(powerconsumption)
 coefs <- learn_noise(df = powerconsumption)
 }
}
}
\seealso{
\code{\link[stats]{smooth.spline}},
\code{\link[stats]{predict}},
\code{\link[stats]{lm}},
\code{\link[stats]{coef}}
}
