% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/helpers.R
\name{lm_error_var}
\alias{lm_error_var}
\title{Calculate error variance given model coefficients.}
\usage{
lm_error_var(...)
}
\arguments{
\item{...}{Pass along all model coefficients, excluding the intercept. These
can be named or unnamed}
}
\value{
Returns the required error variance so that the variance of your
  dependent variable is approximately 1.
}
\description{
\code{lm_error_var} will calculate the required error variance for a linear
model, given specified model coefficients, to create variance for your
dependent variable of approximately 1.
}
\examples{
lm_error_var(b1=.15, b2=.3)  # returns error variance of 0.8875
}

