% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiverse_plot.R
\name{multiverse_plot}
\alias{multiverse_plot}
\title{multiverse_plot}
\usage{
multiverse_plot(
  data,
  outcome = "outcome",
  outcome_name = "Outcome",
  interval_lower = NULL,
  interval_upper = NULL,
  rank_by = "rank",
  rank_by_outcome = FALSE,
  relative_height_of_upper_plot = 0.7,
  outcome_cutoff = NULL,
  limits = NULL
)
}
\arguments{
\item{data}{A data frame containing the simulation parameters, the outcome variable, and the column to rank the rows by.}

\item{outcome}{Column name in the data that represents the outcome variable.}

\item{outcome_name}{Label for the outcome variable in the plot.}

\item{interval_lower}{Column name in the data that represents the lower-bound interval. Defaults to NULL.}

\item{interval_upper}{Column name in the data that represents the upper-bound interval. Defaults to NULL.}

\item{rank_by}{Column name in the data that represents the variable that specifies the ranks order for the outcome (1 = first location on left of the plot).}

\item{rank_by_outcome}{If TRUE, create ranks by ranking the outcome column, therefore ignoring rank_by. Defaults to FALSE.}

\item{relative_height_of_upper_plot}{The height of the upper plot for the simulation outcomes relative to the conditions. Defaults to 0.70. May need to be increased if intervals are used.}

\item{outcome_cutoff}{If a numeric value is specified, a horizontal dashed line will be added to the outcomes plot at this value. This can be used to represent cut-off values, e.g., 80\% power. Defaults to NULL.}
}
\description{
Plot results from a simulation using a multiverse-style plot
}
