% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/helpers.R
\name{n.sims.pwr_sim}
\alias{n.sims.pwr_sim}
\title{Return the number of simulations performed by a power simulation.}
\usage{
\method{n.sims}{pwr_sim}(sim)
}
\arguments{
\item{sim}{A power simulation object of type 'pwr_sim'.}
}
\value{
Returns the number of simulations done in each test.
}
\description{
\code{n.sims.pwr_sim} extracts information about the number of simulations
(per specific test) performed by a power simulation.
}

