% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{n.sims}
\alias{n.sims}
\title{Return the number of simulations performed by a simulation.}
\usage{
n.sims(sim, ...)
}
\arguments{
\item{sim}{A simulation object.}

\item{...}{Additional arguments to be passed to the particular method for the
given object tye.}
}
\value{
The form of the value returned by \code{n.sims} depends on the class
  of its argument. See the documentation of the particular methods for
  details of what is produced by that method.
}
\description{
\code{n.sims} is a generic function that extracts information about the
number of simulations (per specific test) performed by a simulation.
}
\seealso{
\code{\link{n.sims.simulation}}
}

