% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{n.sims.simulation}
\alias{n.sims.simulation}
\title{Return the number of simulations performed by a simulation.}
\usage{
\method{n.sims}{simulation}(sim)
}
\arguments{
\item{sim}{A simulation object of type 'simulation'.}
}
\value{
Returns the number of simulations done in each test.
}
\description{
\code{n.sims.simulation} extracts information about the number of simulations
(per specific test) performed by a simulation.
}

