% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{powerconsumption}
\alias{powerconsumption}
\title{Household Power Consumption}
\format{
A data frame with 745 rows and 1440 variables.
}
\source{
\url{https://archive.ics.uci.edu/ml/datasets/Individual+household+electric+power+consumption}
}
\usage{
powerconsumption
}
\description{
Measurements of electric power consumption in one household
with a one-minute sampling rate over a period of 2 years.
}
\keyword{datasets}
