% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_covariance.R
\name{predict_covariance}
\alias{predict_covariance}
\title{Estimate the covariance given a matrix}
\usage{
predict_covariance(t, covariance)
}
\arguments{
\item{t}{Vector of sampling points.}

\item{covariance}{Matrix of covariance.}
}
\value{
A matrix of covariance estimated on a vector of sampling points.
}
\description{
This function estimates the covariance matrix on a vector of
sampling points. The \code{covariance} matrix parameter has to be larger
than the vector of sampling points.
}
\examples{
\dontrun{
if(interactive()){
 attach(powerconsumption)
 covariance <- learn_covariance(powerconsumption, 'lm')
 cov <- predict_covariance(t = seq(0, 1, length.out = 101),
                           covariance = covariance)
 }
}
}
