% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_mean.R
\name{predict_mean}
\alias{predict_mean}
\title{Estimate the mean function given some model}
\usage{
predict_mean(u, model, lambda, k = 50)
}
\arguments{
\item{u}{PARAM_DESCRIPTION}

\item{model}{PARAM_DESCRIPTION}

\item{lambda}{Value of the penalty parameter}

\item{k}{Number of function to be used to fit the data, default=50}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 attach(powerconsumption)
 mod <- learn_mean(df = powerconsumption, K = 50)
 m <- predict_mean(u = seq(0, 1, length.out = 101), model = mod,
                   lambda = exp(-3.5), k = 50)
 }
}
}
\seealso{
\code{\link[glmnet]{coef.glmnet}}
}
