% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_noise.R
\name{predict_noise}
\alias{predict_noise}
\title{Estimate the variance of the noise given a set of coefficients}
\usage{
predict_noise(t, x, model)
}
\arguments{
\item{t}{Vector of sampling points.}

\item{x}{Vector of observed points.}

\item{model}{Object of class 'gam' from the function \code{learn_noise}.}
}
\value{
A numeric vector representing the estimation of the noise variance.
}
\description{
This function estimates the noise variance given vectors of
sampling and observed points and a vector of coefficients.
}
\examples{
\dontrun{
if(interactive()){
 attach(powerconsumption)
 coefs <- learn_noise(df = powerconsumption)
 noise <- predict_noise(t = seq(0, 1, length.out = 101),
                        x = seq(min(powerconsumption, na.rm = T),
                                max(powerconsumption, na.rm = T),
                                length.out = 101),
                        coefs = coefs)
 }
}
}
