% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pwr_sim.R
\name{pwr_sim}
\alias{pwr_sim}
\title{Simulate statistical power.}
\usage{
pwr_sim(func, params, n.sims = 5000, ...)
}
\arguments{
\item{func}{A user-defined function that generates data and performs a
statistical test on that data.}

\item{params}{A list of parameters to be passed to \code{func}. Each of these
parameters can be a vector of multiple values, and \code{pwr_sim} will test
each combination of these values.}

\item{n.sims}{Number of simulations (per combination of params).}

\item{...}{Additional arguments to be passed to \code{func}. If you do not
need to vary certain parameters in your model, you can pass them to
\code{func} here.}
}
\value{
Returns a data frame with one row per simulation. \code{func} must
  return a (named) vector with the results you wish to capture; these vectors
  will then be combined for the final output.
}
\description{
\code{pwr_sim} simulates data by generating samples according to a user-
defined function. These data can then be used to examine statistical power.
This function is intentionally kept general and flexible, to allow for a wide
variety of statistical models.
}
\examples{
lm_test <- function(N, b0, b1) {
x <- rnorm(N, 0, 1)
y <- rnorm(N, b0 + b1*x, sqrt(1 - b1^2))
data <- data.frame(y, x)
model <- lm(y ~ x, data)

# capture output from model summary
est <- coef(summary(model))['x', 'Estimate']
se <- coef(summary(model))['x', 'Std. Error']
p <- coef(summary(model))['x', 'Pr(>|t|)']

return(c(xm=mean(x), xsd=sd(x), ym=mean(y), ysd=sd(y), est=est, se=se, p=p,
    sig=est > 0 & p <= .05))
}

# test power for two sample sizes: N=200 and N=300, with 5000 sims for each
power <- pwr_sim(lm_test, params=list(N=c(200, 300)), n.sims=5000, b0=0, b1=.15)
}
\seealso{
\code{\link{boot}}
}

