% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/helpers.R
\name{results.pwr_sim}
\alias{results.pwr_sim}
\title{Return results of a power simulation.}
\usage{
\method{results}{pwr_sim}(sim)
}
\arguments{
\item{sim}{A power simulation object of type 'pwr_sim'.}
}
\value{
Returns a data frame with all the data returned from each simulation.
}
\description{
\code{results.pwr_sim} returns the raw data from a power simulation.
}

