% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{results.simulation}
\alias{results.simulation}
\title{Return results of a simulation.}
\usage{
\method{results}{simulation}(sim)
}
\arguments{
\item{sim}{A simulation object of type 'simulation'.}
}
\value{
Returns a data frame with all the data returned from each simulation.
}
\description{
\code{results.simulation} returns the raw data from a simulation.
}

