% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/round_df.R
\name{round_df}
\alias{round_df}
\title{round_df}
\usage{
round_df(x, digits = 3)
}
\arguments{
\item{x}{Data frame or tibble to be rounded.}

\item{digits}{Number of digits to round to.}
}
\value{
A data frame or tibble.
}
\description{
Round all numeric columns in a data frame to a given number of places using the round-half-up method, using \code{"janitor::round_half_up()"}.
}
\examples{
library(ggplot2)

mtcars |>
  tibble::rownames_to_column(var = "automobile") |>
  round_df()

}
