% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_data.R
\name{sim_data}
\alias{sim_data}
\title{Simulated data.}
\usage{
sim_data(nunber_sample, number_covariate)
}
\arguments{
\item{nunber_sample}{number of sample in simulated data.}

\item{number_covariate}{number of covariates.}
}
\value{
A dataframe containing follows:
\item{vari}{ ith control}
\item{vari_square}{ squated ith control}
\item{Y}{ Outcome correlated with var1}
\item{D}{ Treatment correlated with var2}
}
\description{
\code{sim_data} is used to generate simulate estimated with sparse linear. Assume Y = -5D + 25var1_square + u and D = -5var1_square.
}
\examples{
sim <- sim_data(nunber_sample = 1000, number_covariate = 5)

}
\references{
No.
}
