% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_tau.R
\name{sim_tau}
\alias{sim_tau}
\title{Simulated targeting estimation with sparse data.}
\usage{
sim_tau(
  nunber_sample,
  number_covariate,
  number_interation,
  oracle = FALSE,
  LASSO = FALSE,
  RL_forest = FALSE,
  RL = FALSE,
  unpenalty_LASSO = FALSE,
  wrong_simple = FALSE,
  over_parameter = FALSE,
  adaptive_LASSO = FALSE,
  MCP_LASSO = FALSE,
  core = 1
)
}
\arguments{
\item{nunber_sample}{number of sample in simulated data.}

\item{number_covariate}{number of covariates.}

\item{core}{number of core to use function.}
}
\value{
A dataframe containing follows:
\item{est}{ Out-of-sample MSE.}
\item{sample}{ Sample size in simulation data.}
\item{method}{ Prediction method.}
}
\description{
\code{sim_tau} is used to simulate estimated targeting parameter with sparse linear data given N.
}
\examples{
sim <- sim_fix_n(number_sample = 100; number_covariate = 10; core = 1)

}
\references{
No.
}
