% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_decision_rate.R
\name{summarize_decision_rate}
\alias{summarize_decision_rate}
\title{summarize_decision_rate}
\usage{
summarize_decision_rate(nested_results)
}
\arguments{
\item{nested_results}{A nested data frame.}
}
\description{
Summarize the decision based on the p value for Y ~ X across iterations of a fitted analysis. Useful for power analyses.
}
\examples{
# population model
population_model <-
  create_population_model_with_random_item_loadings(
    model_specification = "Y_latent ~ 0*X_latent", # true effect size of zero
    item_loading_min_y = 0.99,
    item_loading_max_y = 0.99,
    n_indicators_y = 8,
  )

# run simulations
results <- 
  generate_data(pop_model_label = "ttest indicators",
                pop_model = population_model, 
                factorial_design = TRUE,
                n = 100, 
                iterations = 1000) |>
  data_processing(method = use_latent_scores) |>
  fit_model(analysis = analysis_ttest)

# decision rate should equal the alpha value of the test when the true effect size is zero (.05)
results |>
  summarize_decision_rate()
 
}
