% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_mean_estimates.R
\name{summarize_mean_estimates}
\alias{summarize_mean_estimates}
\title{summarize_mean_estimates}
\usage{
summarize_mean_estimates(nested_results)
}
\arguments{
\item{nested_results}{A nested data frame.}
}
\description{
Summarize the mean estimates across across iterations of a fitted regression analysis (simple regression, confounder, mediation, collider, or correlational).
}
\examples{
# population model
population_model <-
  create_population_model_with_static_item_loadings(
    model_specification = 
      "Y_latent ~ 0.5*X_latent + 0.5*M_latent; 
       X_latent ~~ 0.5*M_latent",
    item_loading_y = 0.8,
    item_loading_x = 0.7,
    item_loading_m = 0.6,
    n_indicators_y = 10,
    n_indicators_x = 8,
    n_indicators_m = 6
  )

# run a simulation
results <- 
  generate_data(pop_model_label = "covariate indicators",
                pop_model = population_model, 
                n = 100, 
                iterations = 15) |>
  data_preprocessing(method = convert_to_likert) |>
  data_processing(method = calculate_mean_scores) |>
  fit_model(analysis = analysis_correlations) |>
  summarize_mean_estimates()
 
}
