% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.simulation}
\alias{summary.simulation}
\title{Print summary of simulations.}
\usage{
\method{summary}{simulation}(sim, outcome, params = NULL, func = mean, ...)
}
\arguments{
\item{sim}{A list of class 'simulation' generated by the function
\code{simulate}.}

\item{outcome}{A string with the variable name across which to generate the
summaries.}

\item{params}{If NULL, the function will return a summary across all
generated simulation parameters. A character vector can also be included
here, in which case the function will return summaries for only the
selected parameter names. Finally, users can input a list, with the names
of the elements corresponding to parameter names, and the values of each
element providing a vector with the parameter values to be included in the
summary. In all cases, the summary will collapse across all parameters or
parameter values that are excluded from \code{params}.}

\item{func}{The function to be used to generate the summary (e.g., mean, sum).}

\item{...}{Any other arguments to be passed to func.}
}
\value{
Returns a data frame with one row per set of unique simulation tests.
}
\description{
\code{summary.simulation} calculates a summary of a given variable across all,
or a portion of, the simulations.
}
\examples{
# will calculate the mean for all 'sig' values in all simulations where N=200,
# and all simulations where N=300
summary(power_sim, 'sig', params=list(N=c(200, 300)))
}
\seealso{
\code{\link{simulate}}
}

