% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/helpers.R
\name{tests.pwr_sim}
\alias{tests.pwr_sim}
\title{Return the tests performed by a power simulation.}
\usage{
\method{tests}{pwr_sim}(sim)
}
\arguments{
\item{sim}{A power simulation object of type 'pwr_sim'.}
}
\value{
Returns a data frame with one row for each set of simulations that
  was performed.
}
\description{
\code{results.pwr_sim} extracts information about the set of specific tests
(parameter values) for a power simulation.
}

