% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/helpers.R
\name{tests}
\alias{tests}
\title{Return the tests performed by a simulation.}
\usage{
tests(sim, ...)
}
\arguments{
\item{sim}{A simulation object.}

\item{...}{Additional arguments to be passed to the particular method for the
given object tye.}
}
\value{
The form of the value returned by \code{tests} depends on the class
  of its argument. See the documentation of the particular methods for
  details of what is produced by that method.
}
\description{
\code{tests} is a generic function that extracts information about the set of
specific tests (parameter values) for a simulation.
}
\seealso{
\code{\link{tests.simulation}}
}

