% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/helpers.R
\name{tests.simulation}
\alias{tests.simulation}
\title{Return the tests performed by a simulation.}
\usage{
\method{tests}{simulation}(sim)
}
\arguments{
\item{sim}{A simulation object of type 'simulation'.}
}
\value{
Returns a data frame with one row for each set of simulations that
  was performed.
}
\description{
\code{tests.simulation} extracts information about the set of specific tests
(parameter values) for a simulation.
}

