% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{timing.simulation}
\alias{timing.simulation}
\title{Return the timing information of a simulation.}
\usage{
\method{timing}{simulation}(sim)
}
\arguments{
\item{sim}{A simulation object of type 'simulation'.}
}
\value{
Returns an object of class "proc_time" with information about how
  long the simulation process took.
}
\description{
\code{timing.simulation} returns the raw data from a simulation.
}

