% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_latent_scores.R
\name{use_latent_scores}
\alias{use_latent_scores}
\title{use_latent_scores}
\usage{
use_latent_scores(data)
}
\arguments{
\item{data}{A nested data frame.}
}
\value{
A nested data frame of data_raw + new scored data column.
}
\description{
Extract the latent scores already calculated in a previous data generation step.
}
\examples{
# population model
population_model <-
  create_population_model_with_random_item_loadings(
    model_specification = "Y_latent ~ 0.2*X_latent",
    item_loading_min_y = 0.5,
    item_loading_max_y = 0.9,
    n_indicators_y = 8,
  )

# run a simulation
results <- 
  generate_data(pop_model_label = "ttest indicators",
                pop_model = population_model, 
                factorial_design = TRUE,
                n_mean = 100,
                n_sd = 25,
                iterations = 25) |>
  data_processing(method = use_latent_scores)

}
