% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_to_histogram.R
\name{vector_to_histogram}
\alias{vector_to_histogram}
\title{vector_to_histogram}
\usage{
vector_to_histogram(
  simulated_scores,
  binwidth = 0.1,
  xmin = NULL,
  xmax = NULL,
  fill = "#702963"
)
}
\arguments{
\item{simulated_scores}{numeric vector of values to be plotted}

\item{binwidth}{numeric binwidth argument for (\code{"ggplot2::geom_histogram()}.}

\item{xmin}{numeric the xmin argument for (\code{"ggplot2::scale_x_continuous()}, used in both the breaks and limits arguments}

\item{xmax}{numeric the xmax argument for (\code{"ggplot2::scale_x_continuous()}, used in both the breaks and limits arguments}

\item{fill}{string the fill color argument for (\code{"ggplot2::geom_histogram()}}
}
\value{
A ggplot object
}
\description{
#' Plot a numeric vector as a ggplot2 histogram
}
\examples{
library(ggplot2)

rnorm(n = 50, # sample n
      mean = 0, # population mean (μ or mu)
      sd = 1) |> # population sd (σ or sigma)
  vector_to_histogram(xmin = -5, xmax = +5)

rnorm(n = 50, # sample n
      mean = 0, # population mean (μ or mu)
      sd = 1) |> # population sd (σ or sigma)
  vector_to_histogram(binwidth = 1, xmin = -5, xmax = +5)
  
rnorm(n = 50, # sample n
        mean = 0, # population mean (μ or mu)
        sd = 1) |> # population sd (σ or sigma)
  vector_to_histogram()

runif(n = 50, # sample n
      min = -5, # population mean (μ or mu)
      max = +5) |> # population sd (σ or sigma)
  vector_to_histogram(fill = "darkcyan")

rnorm(n = 100000, # sample n
      mean = 0, # population mean (μ or mu)
      sd = 1) |> # population sd (σ or sigma)
  vector_to_histogram(xmin = -5, xmax = +5)

runif(n = 100000, # sample n
      min = -5, # population mean (μ or mu)
      max = +5) |> # population sd (σ or sigma)
  vector_to_histogram(xmin = -5, xmax = +5, fill = "darkcyan") +
  coord_cartesian(ylim = c(0, 4100))

}
