% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_to_histogram_with_summary_statistics.R
\name{vector_to_histogram_with_summary_statistics}
\alias{vector_to_histogram_with_summary_statistics}
\title{vector_to_histogram_with_summary_statistics}
\usage{
vector_to_histogram_with_summary_statistics(
  simulated_scores,
  binwidth = 0.1,
  xmin = NULL,
  xmax = NULL,
  x_prop = 1,
  y_prop = 0.95,
  fill = "#702963"
)
}
\arguments{
\item{simulated_scores}{numeric vector of values to be plotted}

\item{binwidth}{numeric binwidth argument for (\code{"ggplot2::geom_histogram()}.}

\item{xmin}{numeric the xmin argument for (\code{"ggplot2::scale_x_continuous()}, used in both the breaks and limits arguments}

\item{xmax}{numeric the xmax argument for (\code{"ggplot2::scale_x_continuous()}, used in both the breaks and limits arguments}

\item{x_prop}{numeric the x coordinate, in proportions of plot, where the annotation should appear}

\item{y_prop}{numeric the y coordinate, in proportions of plot, where the annotation should appear}

\item{fill}{string the fill color argument for (\code{"ggplot2::geom_histogram()}}
}
\value{
A ggplot object
}
\description{
Plot a numeric vector as a ggplot2 histogram with annotated sample mean and SD
}
\examples{
library(ggplot2)

rnorm(n = 100000, # sample n
      mean = 0, # population mean (μ or mu)
      sd = 1) |> # population sd (σ or sigma)
  vector_to_histogram_with_summary_statistics(xmin = -5, xmax = +5)

runif(n = 100000, # sample n
      min = -5, # population mean (μ or mu)
      max = +5) |> # population sd (σ or sigma)
  vector_to_histogram_with_summary_statistics(xmin = -5, xmax = +5, fill = "darkcyan") +
  coord_cartesian(ylim = c(0, 4100))

}
