% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_correlations.R
\name{analysis_correlations}
\alias{analysis_correlations}
\title{analysis_correlations}
\usage{
analysis_correlations(data)
}
\arguments{
\item{data}{A nested data frame containing the processed data.}
}
\value{
A nested data frame containing the fitted results. The "decision" variable refers to whether a statistically correlation between Y and X (Y ~~ X) was found. This is somewhat arbitrary and could be rethought.
}
\description{
Estimate correlations between X, M, and Y, as well as their 95\% CIs and p values.
}
