% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_ttest.R
\name{analysis_ttest}
\alias{analysis_ttest}
\title{analysis_ttest}
\usage{
analysis_ttest(data)
}
\arguments{
\item{data}{A nested data frame containing the processed data.}
}
\description{
Fit a t test to the data using the model Y ~ X and extract the estimated mean difference, the effects, 95\% CIs and p values.
}
