% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_mean_scores.R
\name{calculate_mean_scores}
\alias{calculate_mean_scores}
\title{calculate_mean_scores}
\usage{
calculate_mean_scores(data)
}
\arguments{
\item{data}{A nested data frame.}
}
\value{
A nested data frame of data_raw + new scored data column.
}
\description{
Convert individual items to a mean score.
}
