% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_to_likert.R
\name{convert_to_likert}
\alias{convert_to_likert}
\title{convert_to_likert}
\usage{
convert_to_likert(data)
}
\arguments{
\item{data}{A nested data frame.}
}
\value{
A nested data frame of data_raw + new preprocessed data column.
}
\description{
Convert continuous raw data to interval Likert-like data. Assumes simulated raw data has mean = 0 and SD = 1. Note that this is violated when generating factorials data, especially when the effect size is very large.
}
