% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_cohens_d_effect_sizes.R
\name{extract_cohens_d_effect_sizes}
\alias{extract_cohens_d_effect_sizes}
\title{summarize_mean_estimates}
\usage{
extract_cohens_d_effect_sizes(nested_fits)
}
\arguments{
\item{nested_fits}{A nested data frame.}
}
\description{
Extract the Cohen's d estimates from each iteration of a fitted t test analysis, e.g., for later meta analysis and/ forest plotting.
}
\details{
Note that this function extracts the \emph{unstandardized} effect size on the assumption that the population model has an SD of 1.0. If this isn't the case
}
