% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_data.R
\name{generate_data}
\alias{generate_data}
\title{generate_data}
\usage{
generate_data(
  pop_model_label,
  pop_model,
  iterations,
  n,
  n_mean,
  n_sd,
  factorial_design = FALSE
)
}
\arguments{
\item{pop_model_label}{string label for the population model, e.g., "confounder indicators" for a confounder causal model with data for individual indicators (e.g., self report items) as well as the latent scores (e.g., sum scores)}

\item{pop_model}{string specifying population model using lavaan notation.}

\item{iterations}{number of iterations aka simulated studies}

\item{n}{sample size per iteration (simulated study). Either n or both n_mean and n_sd must be specified but not both.}

\item{n_mean}{integer mean sample size per iteration (simulated study). Either n or both n_mean and n_sd must be specified but not both. Note that generated values <6 will be changed to 6.}

\item{n_sd}{integer SD sample size per iteration (simulated study). Either n or both n_mean and n_sd must be specified but not both.}

\item{factorial_design}{boolean indicating if data should be generated for a single group cross sectional design or a two group factorial design.}
}
\value{
A nested data frame of data_raw + new scored data column.
}
\description{
Generate simulated data using a given population model, n iterations, and sample size.
}
\details{
To generate both item level data and latent scores, use a helper function (e.g., create_population_model_with_static_item_loadings()
or create_population_model_with_random_item_loadings()) create the model specification which must include the assignment of indicators via "=~".
To generate only the latent variables, specify only the regressions (e.g., "Y ~ 0.5\emph{X + 0.0}M"). See the vignettes for examples.

Note that for the later processing, hacking, and analysis functions to work, generate_data() must produce data with the correct variable names:
To generate item level data and latent variables, you must specify the population model to include variables "Y_latent", "X_latent", and (for some) "M_latent".
To generate just the latent variables/scored data, you must specify the population model to include the variables "Y", "X", and (for some) "M" (i.e., no "_latent").
}
