% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_data.R
\name{generate_data}
\alias{generate_data}
\title{FUNCTION_TITLE}
\usage{
generate_data(n, m, model_mean, covariance, coefs, lambda, p)
}
\arguments{
\item{n}{PARAM_DESCRIPTION}

\item{m}{PARAM_DESCRIPTION}

\item{model_mean}{PARAM_DESCRIPTION}

\item{covariance}{PARAM_DESCRIPTION}

\item{coefs}{PARAM_DESCRIPTION}

\item{lambda}{PARAM_DESCRIPTION}

\item{p}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 attach(powerconsumption)
 mod <- learn_mean(df = powerconsumption, k = 50)
 cov <- learn_covariance(powerconsumption, 'lm')
 coefs <- learn_noise(df = powerconsumption)
 df <- generate_data(n = 10, m = 40, model_mean = mod, covariance = cov,
                     coefs = coefs, lambda = exp(-3.5), p = 0.2)
 }
}
}
\seealso{
\code{\link[purrr]{map}},\code{\link[purrr]{map2}}
\code{\link[stats]{Uniform}},\code{\link[stats]{Normal}}
\code{\link[MASS]{mvrnorm}}
}
