% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_covariance.R
\name{learn_covariance}
\alias{learn_covariance}
\title{FUNCTION_TITLE}
\usage{
learn_covariance(df, method = "lm")
}
\arguments{
\item{df}{PARAM_DESCRIPTION}

\item{method}{PARAM_DESCRIPTION, default='lm'}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 attach(powerconsumption)

 # Using 'lm'
 cov <- learn_covariance(powerconsumption, 'lm')

 # Using 'min'
 cov <- learn_covariance(powerconsumption, 'min')
 }
}
}
\seealso{
\code{\link[stats]{smooth.spline}},
\code{\link[stats]{predict}},
\code{\link[fdapace]{Lwls2D}}
}
