% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_mean.R
\name{learn_mean}
\alias{learn_mean}
\title{Create a model for the estimation of the mean}
\usage{
learn_mean(df, k = 50)
}
\arguments{
\item{df}{PARAM_DESCRIPTION}

\item{k}{Number of function to be used to fit the data, default=50.}
}
\value{
A \code{\link[glmnet]{glmnet}} model
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 attach(powerconsumption)
 mod <- learn_mean(df = powerconsumption, k = 50)
 }
}
}
\seealso{
\code{\link[glmnet]{glmnet}}
}
