% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_noise.R
\name{learn_noise}
\alias{learn_noise}
\title{Estimate a set of coefficients for the noise variance}
\usage{
learn_noise(df)
}
\arguments{
\item{df}{Dataframe containing the real dataset.}
}
\value{
An object of class 'gam' from the function \code{mgcv::gam}.
}
\description{
This function estimates the noise variance by fitting a linear
model on the squared difference between the observed curves and their
smoothed version.
}
\examples{
\dontrun{
if(interactive()){
 attach(powerconsumption)
 coefs <- learn_noise(df = powerconsumption)
 }
}
}
\seealso{
\code{\link[stats]{smooth.spline}},
\code{\link[stats]{predict}},
\code{\link[stats]{lm}},
\code{\link[stats]{coef}}
}
