% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaanalysis.R
\name{metaanalysis}
\alias{metaanalysis}
\title{metaanalysis}
\usage{
metaanalysis(results, published_only = FALSE)
}
\arguments{
\item{results}{Nested data frame containing results, as outputted by \code{"fit_model()"}.}
}
\description{
Fit a random effects meta-analysis to the results across iterations.
}
\examples{
set.seed(42)

# population model
population_model <-
  create_population_model_with_random_item_loadings(
    model_specification = "Y_latent ~ 0.2*X_latent",
    item_loading_min_y = 0.5,
    item_loading_max_y = 0.9,
    n_indicators_y = 8,
  )

# run a simulation
results <- 
  generate_data(pop_model_label = "ttest indicators",
                pop_model = population_model, 
                factorial_design = TRUE,
                #n = 50, 
                n_mean = 100,
                n_sd = 25,
                iterations = 25) |>
  data_preprocessing(method = convert_to_likert) |>
  data_processing(method = use_latent_scores) |>
  fit_model(analysis = analysis_ttest) |>
  extract_cohens_d_effect_sizes()

# fit meta analysis across iterations
metaanalysis(results)
 
}
