% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_covariance.R
\name{predict_covariance}
\alias{predict_covariance}
\title{FUNCTION_TITLE}
\usage{
predict_covariance(t, covariance)
}
\arguments{
\item{t}{PARAM_DESCRIPTION}

\item{covariance}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 attach(powerconsumption)
 covariance <- learn_covariance(powerconsumption, 'lm')
 cov <- predict_covariance(t = seq(0, 1, length.out = 101),
                           covariance = covariance)
 }
}
}
