% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_mean.R
\name{predict_mean}
\alias{predict_mean}
\title{Estimate the mean function given some model}
\usage{
predict_mean(u, model, lambda, k = 50)
}
\arguments{
\item{u}{Vector of sampling points.}

\item{model}{\code{\link[glmnet]{glmnet}} model for the mean curve.}

\item{lambda}{Value of the penalty parameter.}

\item{k}{Number of function to be used to fit the data, default=50.}
}
\value{
A numeric vector representing the estimation of the mean curve.
}
\description{
This function estimates the mean curve of a functional dataset
given a vector of sampling points and a \code{\link[glmnet]{glmnet}} model
for the mean curve.
}
\examples{
\dontrun{
if(interactive()){
 attach(powerconsumption)
 mod <- learn_mean(df = powerconsumption, K = 50)
 m <- predict_mean(u = seq(0, 1, length.out = 101), model = mod,
                   lambda = exp(-3.5), k = 50)
 }
}
}
\seealso{
\code{\link[glmnet]{coef.glmnet}}
}
