% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_noise.R
\name{predict_noise}
\alias{predict_noise}
\title{FUNCTION_TITLE}
\usage{
predict_noise(t, x, coefs)
}
\arguments{
\item{t}{PARAM_DESCRIPTION}

\item{x}{PARAM_DESCRIPTION}

\item{coefs}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 attach(powerconsumption)
 coefs <- learn_noise(df = powerconsumption)
 noise <- predict_noise(t = seq(0, 1, length.out = 101),
                        x = seq(min(powerconsumption, na.rm = T),
                                max(powerconsumption, na.rm = T),
                                length.out = 101),
                        coefs = coefs)
 }
}
}
