% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_decision_rate.R
\name{summarize_decision_rate}
\alias{summarize_decision_rate}
\title{summarize_decision_rate}
\usage{
summarize_decision_rate(nested_results)
}
\arguments{
\item{nested_results}{A nested data frame.}
}
\description{
Summarize the decision based on the p value for Y ~ X across iterations of a fitted analysis. Useful for power analyses.
}
