% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_mean_estimates.R
\name{summarize_mean_estimates}
\alias{summarize_mean_estimates}
\title{summarize_mean_estimates}
\usage{
summarize_mean_estimates(nested_results)
}
\arguments{
\item{nested_results}{A nested data frame.}
}
\description{
Summarize the mean estimates across across iterations of a fitted regression analysis (simple regression, confounder, mediation, collider, or correlational).
}
