% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_latent_scores.R
\name{use_latent_scores}
\alias{use_latent_scores}
\title{use_latent_scores}
\usage{
use_latent_scores(data)
}
\arguments{
\item{data}{A nested data frame.}
}
\value{
A nested data frame of data_raw + new scored data column.
}
\description{
Extract the latent scores already calculated in a previous data generation step.
}
